chore(copier): heavy backfill v1.0.0 -> v1.1.0 + bootstrap copier-update workflow
Bootstraps .github/workflows/copier-update.yml (the new weekly auto
template-sync workflow shipped in template v1.1.0), advances
.copier-answers.yml _commit: to v1.1.0, and restructures hybrid
files so future copier update runs do not silently overwrite
project-owned divergences.

Heavy-backfill changes:

- pyproject.toml: license + MIT classifier divergence comments;
  PROJECT-DEPS-START/END + PROJECT-EXTRAS-START/END sentinel zones.
- packaging/nfpm.yaml: license divergence comment.
- src/markdown_vault_mcp/cli.py -> 3-line shim that re-exports
  main from _cli_impl.py (project-owned, not in template).
- tests/test_cli.py: imports updated to point at _cli_impl.
